refactor(filid): remove Config Proposal Discipline from agent documentation
- Removed the `## Config Proposal Discipline` section from the adjudicator, engineering-architect, knowledge-manager, and operations-sre documentation files. This section previously outlined requirements for proposing or evaluating `.filid/config.json` patches, including schema citation and handling unverifiable config keys.
- The removal aligns with the recent restructuring of the configuration validation process and focuses the documentation on current practices without outdated references.
